Update auth entity 제구성
